Make unit tests pass under Python 3.
Most of the issues were related to text encoding (byte vs utf-8), and XML
attributes ordering. These were fixed by modifying the tests to work with both
Python 3, and Python 2.7.

One interesting library issue identified was the ordering of elements in the XML
document that is produced from PyACI. As it was internally using dictionary to
store the list of children, the resulting element order was at the mercy of the
order of dictionary keys. To make this work correctly, modified the library to
use OrderedDict.